Ex5/requirements.txt
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
db-dtypes==1.2.0
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.8.0
google-api-core==2.19.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-storage
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pandas-gbq==0.23.1
proto-plus==1.24.0
protobuf==5.27.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydata-google-auth==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==71.0.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
storage